+++ src/main/java/com/codeup/on_the_house/security/OAuthConfiguration.java
@@ -22,7 +22,7 @@ public class OAuthConfiguration extends AuthorizationServerConfigurerAdapter {
 
     private final UserDetailsService userService;
 
-    @Value("${jwt.clientId:rest-blog-client}")
+    @Value("${jwt.clientId:on-the-house-client}")
     private String clientId;
 
     @Value("${jwt.client-secret:secret}")
