+++ src/main/java/com/codeup/on_the_house/security/ResourceServerConfiguration.java
@@ -33,13 +33,19 @@ public class ResourceServerConfiguration extends ResourceServerConfigurerAdapter
                 .sessionCreationPolicy(SessionCreationPolicy.STATELESS)
                 .and()
                 .authorizeRequests()
-//                .antMatchers("/api/users")
+                .antMatchers("/api/users", "api/users/**")
+                .hasAnyAuthority("ADMIN", "USER")
+                .antMatchers("/api/posts", "api/posts/**")
+                .hasAnyAuthority("ADMIN", "USER")
+//                .antMatchers("/home")
 //                .hasAnyAuthority("ADMIN", "USER")
-//                .antMatchers("/api/posts")
+//                .antMatchers("/profile")
+//                .hasAnyAuthority("ADMIN", "USER")
+//                .antMatchers("/events")
 //                .hasAnyAuthority("ADMIN", "USER")
                 .antMatchers("/swagger-ui/**", "/v3/api-docs/**")
                 .permitAll()
-                .antMatchers("/api/users/create")
+                .antMatchers("/api/users/createUser")
                 .permitAll()
                 .antMatchers("/**")
                 .permitAll()
