+++ src/main/java/com/codeup/on_the_house/service/PostService.java
@@ -12,6 +12,8 @@ import java.util.List;
 
 @Service
 public class PostService {
+
+
     private final PostsRepository postsRepository;
     private final UserService userService;
