+++ src/main/resources/static/js/createView.js
@@ -26,6 +26,7 @@ export default function createView(URI) {
     let request = {
         headers: getHeaders()
     }
+    console.log(`Route: ` + route +  `\n` + "Request: " + request);
     fetchData(route.state, request).then((props) => {
         // Restore the title so that history entries are not all 'Loading...'
         // I tried using route.uri here instead, but it seems there's an off-by-one bug (https://stackoverflow.com/a/38830794)
@@ -33,7 +34,8 @@ export default function createView(URI) {
         // Add the current page to the history stack
         history.pushState({...props, lastUri: route.uri }, null, route.uri)
         render(props, route);
-    });
+    })
+        // .catch(err => {console.log(err); createView("/")});
 }
 
 // When the user hits back in the browser, get the last uri from history and render it (w/ props)
