+++ src/main/resources/static/js/views/partials/postCards.js
@@ -55,10 +55,43 @@ export function markerPostCards(user) {
     return htmlCard
 }
 
-export function myPostCards(post) {
-    //language=HTML
-    console.log(post)
+// export function myPostCards(post) {
+//     //language=HTML
+//     console.log(post)
+//
+//     let html = ``
+//
+//     for (let i = 0; i < post.length; i++) {
+//         html += `
+//          <div class="card mb-3 col-3 mx-2" style="max-width: 550px; height: 475px border-radius: 2%; box-shadow: 0 4px 8px 0 rgba(0, 0, 0, 0.2), 0 6px 20px 0 rgba(0, 0, 0, 0.19);">
+//     <div>
+//       <div class="card-body">
+//         <div class="d-flex justify-content-between align-items-center">
+//             <h5 class="card-title">${post[i].itemName}</h5>
+//         </div>
+//         <img src='https://images.pexels.com/photos/235294/pexels-photo-235294.jpeg?auto=compress&cs=tinysrgb&w=1260&h=750&dpr=2' class="img-fluid rounded-start my-2" alt="..." id="card-img-home">
+//         <p class="card-text" style="height: 40px">${post[i].description}</p>
+//         <p class="card-text"><small>Quantity: ${post[i].quantity}</small></p>
+//         <p class="card-text"><small>Expiry Date: ${post[i].expiryDate}</small></p>
+//         <p class="card-text"><small>Shared by: ${post[i].user.username}</small></p>
+//       </div>
+//       <div class="d-flex mb-2 justify-content-around">
+//               ${editPostModal(post[i])}
+//               <button id="deleteBtn" type="button" class="btn btn-danger" onclick="confirmAction()" data-id="${post[i].id}">Delete</button>
+//        </div>
+//     </div>
+// </div>
+//
+// <br>
+// `
+//     }
+//     return html
+//
+// }
 
+export function myPostCards(user) {
+    //language=HTML
+    let post = user.posts;
     let html = ``
 
     for (let i = 0; i < post.length; i++) {
@@ -73,7 +106,7 @@ export function myPostCards(post) {
         <p class="card-text" style="height: 40px">${post[i].description}</p>
         <p class="card-text"><small>Quantity: ${post[i].quantity}</small></p>
         <p class="card-text"><small>Expiry Date: ${post[i].expiryDate}</small></p>
-        <p class="card-text"><small>Shared by: ${post[i].user.username}</small></p>
+        <p class="card-text"><small>Shared by: ${user.username}</small></p>
       </div>
       <div class="d-flex mb-2 justify-content-around">
               ${editPostModal(post[i])}
